+++ frontend/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MessageListActivity"
+            android:name=".view.MessageListActivity"
             android:label="@string/title_message_list"
             android:parentActivityName=".view.MainActivity"
             android:theme="@style/AppTheme.NoActionBar">
@@ -25,13 +25,13 @@
                 android:value="com.wordpress.commonplayground.view.MainActivity" />
         </activity>
         <activity
-            android:name=".MessageDetailActivity"
+            android:name=".view.MessageDetailActivity"
             android:label="@string/title_message_detail"
-            android:parentActivityName=".MessageListActivity"
+            android:parentActivityName=".view.MessageListActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.wordpress.commonplayground.MessageListActivity" />
+                android:value="com.wordpress.commonplayground.view.MessageListActivity" />
         </activity>
         <activity
             android:name=".view.MainActivity"
