+++ frontend/app/src/main/java/com/wordpress/commonplayground/view/MessagesAdapter.java
@@ -1,0 +1,62 @@
+package com.wordpress.commonplayground.view;
+
+import android.content.Context;
+import android.content.Intent;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import com.wordpress.commonplayground.R;
+import com.wordpress.commonplayground.model.Message;
+
+import java.util.List;
+
+public class MessagesAdapter extends RecyclerView.Adapter<MessagesAdapter.ViewHolder> {
+        private final List<Message> mValues;
+
+        private final View.OnClickListener mOnClickListener = new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Message item = (Message) view.getTag();
+                Context context = view.getContext();
+                Intent intent = new Intent(context, MessageDetailActivity.class);
+                intent.putExtra(MessageDetailFragment.ARG_ITEM_ID, item.getDescription());
+
+                context.startActivity(intent);
+            }
+        };
+
+        MessagesAdapter(List<Message> items) {
+            mValues = items;
+        }
+
+        @Override
+        public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+            View view = LayoutInflater.from(parent.getContext())
+                    .inflate(R.layout.message_list_content, parent, false);
+            return new ViewHolder(view);
+        }
+
+        @Override
+        public void onBindViewHolder(final ViewHolder holder, int position) {
+            holder.mContentView.setText(mValues.get(position).getTitle());
+            holder.itemView.setTag(mValues.get(position));
+            holder.itemView.setOnClickListener(mOnClickListener);
+        }
+
+        @Override
+        public int getItemCount() {
+            return mValues.size();
+        }
+
+        class ViewHolder extends RecyclerView.ViewHolder {
+            final TextView mContentView;
+
+            ViewHolder(View view) {
+                super(view);
+                mContentView = (TextView) view.findViewById(R.id.content);
+            }
+        }
+    }
