+++ frontend/app/src/main/res/layout/activity_message_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".MessageListActivity">
+    tools:context=".view.MessageListActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar"
